+++ src/common/Ship.js
@@ -34,7 +34,7 @@ export default class Ship extends PhysicalObject2D {
                 type: BaseTypes.TYPES.LIST,
                 itemType: BaseTypes.TYPES.UINT8 // 0=unloaded,n=torp type
             },
-            damage: { type: BaseTypes.TYPES.INT32 },
+            damage: { type: BaseTypes.TYPES.INT16 },
             pdcAngle: { type: BaseTypes.TYPES.FLOAT32 },
             pdcState: { type: BaseTypes.TYPES.UINT8 } //0=off,1=active,2=firing
         }, super.netScheme);
